Rename `_app+` folder to `_protected+`
