+++ frontend/app/routes/_protected+/_index.tsx
