+++ frontend/app/routes/_protected+/_layout.tsx
